dir restructure + import as types
