+++ src/ui/components/Referenda.tsx
@@ -1,15 +1,16 @@
-import { memo, useState } from 'react';
-import { Remark } from 'react-remark';
-import { useSprings, animated } from '@react-spring/web';
-import { useDrag } from '@use-gesture/react';
-import { createStandardAccountVote } from '../chain/conviction-voting';
-import { Card, Loading } from '../ui/nextui';
-import {
+import type {
   AccountVote,
   ReferendumDetails,
   ReferendumOngoing,
   Track,
-} from '../types';
+} from '../../types';
+
+import { memo, useState } from 'react';
+import { Remark } from 'react-remark';
+import { useSprings, animated } from '@react-spring/web';
+import { useDrag } from '@use-gesture/react';
+import { createStandardAccountVote } from '../../chain/conviction-voting';
+import { Card, Loading } from '../lib';
 
 function Header({
   index,
