+++ src/ui/lib/Icons.tsx
@@ -1,0 +1,3 @@
+import Heart from 'jsx:../icons/heart.svg';
+import Square from 'jsx:../icons/square.svg';
+export { Heart as HeartIcon, Square as CloseSquareIcon };
