--- src/ui/nextui/Icons.tsx
@@ -1,3 +1,0 @@
-import Heart from 'jsx:./icons/heart.svg';
-import Square from 'jsx:./icons/square.svg';
-export { Heart as HeartIcon, Square as CloseSquareIcon };
